Add EA & NE public face area lookup service
This change adds the first of 2 area lookups; the Environment Agency and Natural England Public Face Area lookup.

As its the first it also includes some general purpose classes, namely the base service class and the reponse object.

When we make a request we want to always respond with 2 values; whether the lookup was successful, and the area if it was else an error if it wasn't.

As such we create a response object to retain all values, which a calling app can then use.

Using a response object we feel is an improvement over previous implementations, which relied on the calling app to determine if the request was successful based on whether a string or an error was returned.

Now they will know its always a response object returned, and 'it' will confirm whether the request was successful.