Fix broken public face area service
Whilst updating the README (coming in a future commit) we ran the query for the public face service manully using CURL and got no response.

When implementing the Water Management service we initially just copied the public face url and updated the dataset within it. However this didn't work. We then went back to the working example provided by the Data Systems team and spotted that the version had been dropped to 1.0.0 for Water Management.

Using that got the Water Management query working. However today we tried running the public face service manually and again got no result. So on the off chance we tried changing the version to 1.0.0 and voila, we have a result!

Finally changing the version seems to change the format of the result, hence we also had to tweak how we parse the xml 😩!