Add timeout configuration option
By default the timeout for a [rest-client](https://github.com/rest-client/rest-client#timeouts) request is 60 seconds.

Due to the known 'flaky' nature of the WFS's were calling we don't want users in the service left hanging for 60 seconds whilst the calling app waits to find out what area a registration belongs to.

So this change adds the timeout option to our rest-client calls, along with a configuration option that host apps can use to override this value.