Remove unnecessary rescue in base_service (#4)
Prior to this change we were rescuing the error when calling the WFS, but then raising it again.

We think it was because of a misplaced belief that without it the `DefraRuby::Area::Response` object would then not get the error to save within the response object. A review of the code and a quick test though confirmed this is rubbish.

Hence this change to remove it.