Refactor type_name to own method
As part of breaking down we are going to need the type name to form our url. But we also use this in the `response_xml_path()` which is used to locate and extract the return value we want from the response from the WFS.

Plus after a quick review type name is the thing that changes between the services, not the actual response path.

Hence this quick refactoring of how `response_xml_path()` works by moving it to the `BaseService` and then creating a method called `type_name()` that must be implemented in the parent services.