Update query to use `typeName' vs 'typeNames'
According to the reference I am referring to version 1.1 and below of the WFS spec states that `typeName` is used. From version 2.0 `typeNames` is used.

The original version of the working example queries we were provided did use version 2.0, but then an update for the Water Management one dropped to 1.0. Next Public areas stopped working and we found switching it also to version 1.0 got it working again.

As the reference states the `typeName` should be used for versions sub 2.0 we're making this change to prevent any confusion in the future.