Make url return values consistent
For type_name we were only returning the actual value, leaving the query parameter name in the url. For the rest we were returning both the argument name and its value.

We want to use the type name in `response_xml_path()` so we cannot return the parameter name in the value. So this makes the return values inconsistent.

Also we think keeping the argument names in place in the url string we think helps with understanding how its made up.

So this change tweaks the return values to be consistent, returning only the values that will be passed to the url query params.