Assume `ms:` prefix will be pre-applied
Prior to this request we handled the `ms:` prefix at the beginning of the typename parameter value.

However we are about to implement a change to improve the readability of the `url()` method, and that change relies on the params being a simple `key=value` pairing.

So this change moves the `ms:` prefix into the actual value defined in the service classes. We also think this is more consistent with the rest of how the params work.