Extract code and short name from WFS's
When we reviewed our existing services they only refer to the long name return by each of the administrative boundary Web Feature Services.

However when we initially went to implement this new gem with the [flood-risk-engine](https://github.com/Defra/flood-risk-engin) we found that it doesn't just update the registration with the long name. It instead creates a record (if one doesn't exist) in a separate table which it then links to the registration.

The object that gets saved to the table is expecting a code, short name and long name. We wanted to avoid doing too much refactoring in the flood-risk-engine as it has not gone through an alignment exercise.

So we've come back to this gem and updated how it works. Now `area` on `DefraRuby::Area::Response` will return an instance of `DefraRuby::Area::Area, which itself contains the code, short name and long name for an administrative area. This means we've also tweaked the query we send to the WFS's to include the additional parameters.