Refactor Area to parse the WFS XML response (#8)
Having added the property `area_id` to the `DefraRuby::Area::Area` class we've hit the general rule that "pass no more than four parameters into a method". We know we need to add `area_name` in our next change, at which point we'll break the rule and rubocop will start to complain.

We could override rubocop as it would just be this one instance, but we decided instead to refactor as per [Preserve Whole Object](https://refactoring.guru/preserve-whole-object).

The [rule comes from Sandi Metz](https://thoughtbot.com/blog/sandi-metz-rules-for-developers) author of **Practical Object-Oriented Design in Ruby**. She's also stated that "you can’t just make it one big hash."

The **perservce whole object refactor** is

> **Problem** You get several values from an object and then pass them as parameters to a method.
> **Solution** Instead, try passing the whole object.

So rather than have the `BaseService` parse the xml for the values and pass them to the `Area` instance, we've refactored it such that the parsing of the XML is now handled by the `Area` class. The `BaseService` just passes what the `Area` instance needs to do this.

With this change we'll be ready to handle adding the area name as an attribute.