Update response to return empty areas array
A great suggestion in the PR feedback was that when we have an error or no matches, rather than `response.areas` being `nil` it should be an empty array.

This is more meaningful and means the host application can rely on always getting an array, rather than having to check for `nil` or an array.