Add support for multiple WFS results (#10)
https://eaflood.atlassian.net/browse/RUBY-484

When we originally integrated with the WFS's we were led to believe that because we were dealing with geospatial shapes that for any one query (coordinate) we would only get one matching administrative boundary.

However our awesome QA @andrewhick was able to find cases where a coordinate would return 2 boundaries.

Initially on the basis the functionality included in this gem is based on previous iterations which have been live for some years we were not intending to change the way things function. As such they would still continue to return the first result listed in the XML response.

In practise because of the way the XML parsing works in the gem, what actually happens is the values are getting concatenated. For example querying 456330 & 267000 returns **Lincolnshire and Northamptonshire** and **Staffordshire Warwickshire and West Midlands**. The resulting `long_name` for the area then becomes `"Lincolnshire and NorthamptonshireStaffordshire Warwickshire and West Midlands"`.

This happens for the all the values which is not acceptable. So we needed to update the way the XML parsing works to resolve it. And as we are refactoring that, we decided we may as well refactor the `DefraRuby::Area::Response` to return an array of all the results, not just the first. The calling app can then determine whether to just use the first result, or do something about using all the results.

Finally as we are now returning an array, we have made it so that in the event of an error or no match we return an empty array rather than `null`. This means the calling app knows it will always get an array, it's just whether it's populated or not.